requirements.txt
sentence-transformers==2.2.2
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-google-genai==0.0.6
langchain-experimental==0.0.55
langdetect==1.0.9
langsmith==0.1.31
numpy
pandas
protobuf==4.25.2
pydantic==2.6.0
python-dotenv==1.0.0
streamlit==1.31.0
python_magic_bin-0.4.14-py2.py3-none-win_amd64.whl
InstructorEmbedding==1.0.1
google-ai-generativelanguage==0.4.0
google-api-core==2.16.1
google-auth==2.27.0
google-generativeai==0.3.2
faiss-cpu==1.7.4
pyarrow==15.0.0
jsonlines==4.0.0
tiktoken==0.4.0
pymysql
chromadb
youtube-transcript-api
pathlib
sentence-transformers==2.2.2
unstructured==0.12.2
InstructorEmbedding